+++ day10/solution.h
@@ -9,8 +9,8 @@
 namespace Day10 {
   auto first_illegal_character(std::string const&) -> std::optional<char>;
   auto calculate_points_for_part1(std::vector<std::string> const &input) -> int;
-  auto mathematical_sum_of_missing_characters(std::string const& line) -> long long int;
-  auto calculate_points_for_part2(std::vector<std::string> const &input) -> long long int;
+  auto mathematical_sum_of_missing_characters(std::string const& line) -> size_t;
+  auto calculate_points_for_part2(std::vector<std::string> const &input) -> size_t;
 }
 
 namespace AoC {
@@ -19,7 +19,7 @@ namespace AoC {
   struct Solution<10> {
     explicit Solution(std::istream &);
     auto part1() const -> int;
-    auto part2() const -> long long int;
+    auto part2() const -> size_t;
 
   private:
     std::vector<std::string> report;
